fix(dashboard): extract shared getBudgetSettings — single source of truth
Argus found getBudgetSettings() duplicated in CostPage and BudgetAlertBanner
with different implementations. Now extracted to utils/budgetSettings.ts:

- Single getBudgetSettings() with Number()/Boolean() coercion
- Single STORAGE_KEY constant ('aegis:settings')
- Both CostPage and BudgetAlertBanner import from shared module
- No more duplicate logic divergence risk

1265 tests pass, tsc clean, build clean.

Refs: #3183 (review)
